+++ src/app/ngx-mask/test/repeat-mask.spec.ts
@@ -1,0 +1,37 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { NgxMaskModule } from '../ngx-mask.module';
+import { ReactiveFormsModule } from '@angular/forms';
+import { TestMaskComponent } from './utils/test-component.component';
+import { equal } from './utils/test-functions.component';
+
+
+describe('Directive: Mask', () => {
+
+  let fixture: ComponentFixture<TestMaskComponent>;
+  let component: TestMaskComponent;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      declarations: [TestMaskComponent],
+      imports: [ReactiveFormsModule, NgxMaskModule.forRoot()]
+    });
+    fixture = TestBed.createComponent(TestMaskComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('repeat mask', () => {
+    component.mask = '0{4}';
+    equal('1234', '1234', fixture);
+  });
+
+  it('repeat mask date', () => {
+    component.mask = '0{2}/0{2}/0{4}';
+    equal('12345678', '12/34/5678', fixture);
+  });
+
+  it('specialCharacters quotes', () => {
+    component.mask = '\'00\'00\'0000\'';
+    equal('12345678', '\'12\'34\'5678\'', fixture);
+  });
+});