+++ CHANGELOG.md
@@ -1,5 +1,17 @@
+<a name="6.5.16"></a>
+# [6.5.16 new specialCharacters 'quotes'] (2018-10-09)
+You can used new specialCharacters 'quotes'
+
+<a name="6.5.15"></a>
+# [6.5.15 support for repeat mask] (2018-10-08)
+You can pass into mask pattern with brackets
+
+<a name="6.4.14"></a>
+# [6.4.14 fixed '*' pattern in maskExpression ] (2018-10-05)
+Fixed pattern '*'
+
 <a name="6.4.13"></a>
-# [6.4.13 selection of input value and deletion of it is fixed ] (2018-04-10)
+# [6.4.13 selection of input value and deletion of it is fixed ] (2018-10-05)
 You can select and delete value.
 
 <a name="6.4.12"></a>
@@ -15,7 +27,7 @@ Fixed when page is loaded the mask is visible
 Fixed writeValue in directive
 
 <a name="6.2.8"></a>
-# [6.2.6 support for Mask Pipe/MaskService] (2018-01-10)
+# [6.2.8 support for Mask Pipe/MaskService] (2018-01-10)
 Added ability to use MaskPipe/MaskService inside component
 
 <a name="6.2.5"></a>